feat(accountDelete): delete user data endpoint
Adds a REST endpoint to delete user data. This clears
data from two sources: highlights (user_annotations) table
and the notes dynamodb.

An additional index was added to the notes dynamo so that
we can query by userId, since userId was not previously
on the table (and we would not know what to delete if the
user_annotations data was already gone).

Moves the isPremium check from NotesDataService to the
Context. Resolvers now use the Context as a factory for
NotesDataService, and the check is lazily performed when
newing up the service. This way we can reuse the
NotesDataService class without requiring premium status,
since we want to be able to delete data from users who
may have once been premium, and may have data in notes.

[INFRA-501]