Infra 501/delete (#207)
* feat(accountDelete): restructure event code

Use conditional lambda deploy pattern

Update environment for lambda

* feat(accountDelete): add userId index to notes dynamodb table

* feat(accountDelete): delete user data endpoint

Adds a REST endpoint to delete user data. This clears
data from two sources: highlights (user_annotations) table
and the notes dynamodb.

An additional index was added to the notes dynamo so that
we can query by userId, since userId was not previously
on the table (and we would not know what to delete if the
user_annotations data was already gone).

Moves the isPremium check from NotesDataService to the
Context. Resolvers now use the Context as a factory for
NotesDataService, and the check is lazily performed when
newing up the service. This way we can reuse the
NotesDataService class without requiring premium status,
since we want to be able to delete data from users who
may have once been premium, and may have data in notes.

[INFRA-501]

* trigger ci

* chore(ci): fix a conditional 🤦

* chore(ci): will it blend

* chore: fix spec test constructor mock

* chore: lint

* chore: lint

* chore: fix typo in service uri

* fix: actually use the route

* Update src/main.ts

Co-authored-by: Kelvin Yeboah <kelvinkyeboah@gmail.com>

* fix(ci): don't try to interpolate env vars

CircleCI won't let you: https://circleci.com/docs/2.0/env-vars/#using-parameters-and-bash-environment

* chore: more literals to fix

* chore(ci): fix capitalization

* chore(ci): fix application codeploy name

* chore(ci): fix bucket name...

Co-authored-by: Kelvin Yeboah <kelvinkyeboah@gmail.com>